# Improve memory efficiency, fixes #2