It is done by dividing each interval into n buckets and keeping a counter in each bucket. Now each (inteval / n) period we increment index of current bucket, looping at the end. This results in RateCounter incrementing in real-time and decrementing every (interval / n).

This way can keep approximate rate and avg rate in constant memory in contrast to previous approach (each increment spawned a goroutine). We can manipulate resolution of RateCounter with nice default of 20 (interval is divided into 20 parts).

Benchmarks before change:

BenchmarkAvgRateCounter-4                     	 3000000	       646 ns/op	      96 B/op	       2 allocs/op
BenchmarkCounter-4                            	200000000	         8.89 ns/op	       0 B/op	       0 allocs/op
BenchmarkRateCounter-4                        	 2000000	       672 ns/op	      96 B/op	       2 allocs/op
BenchmarkRateCounter_Parallel-4               	 3000000	       476 ns/op	     130 B/op	       2 allocs/op
BenchmarkRateCounter_With5MillionExisting-4   	10000000	       236 ns/op	     136 B/op	       2 allocs/op

Benchmarks after change:
BenchmarkAvgRateCounter-4                     	30000000	        56.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkCounter-4                            	200000000	         8.93 ns/op	       0 B/op	       0 allocs/op
BenchmarkRateCounter-4                        	50000000	        27.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkRateCounter_Parallel-4               	20000000	        70.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkRateCounter_With5MillionExisting-4   	50000000	        27.4 ns/op	       0 B/op	       0 allocs/op

As you can see both speed and memory efficiency improved (reaching almost 0 bytes allocates per operation). The amount of memory required is just just O(resolution).

This should fix #2