@sheerun 2017-06-16
I've also added Hits() method to AvgRateCounter to read underlying RateCounter. This way one can improve efficiency even more by using AvgRateCounter for measuring both latencies and rate count.