@paulbellamy 2017-06-16
Thanks! This looks fantastic.

I hope it is be possible to keep the "higher resolution" that individual timers offer while having "minimal" allocations/op. For that, I was planning on trying a `[]time.Time` as a queue with a single goroutine on it. But, the way you've done it gives us 0 allocs/op, which is *really* appealing. Let me have a think about it and try a couple other benchmarks. 🤔