@sheerun 2017-06-16
That was my initial approach as well, but I figured out O(1) is better than O(n).

I could improve algorithm by smoothing the decrementing of last bucket, but it would make algorithm more complex and error-prone (I tried it), and the same effect can be achieved by increasing resolution a little.